The inspector works under windows!
// Warning: there is an issue with STL containers and dlls!
// See https://stackoverflow.com/questions/4145605/stdvector-needs-to-have-dll-interface-to-be-used-by-clients-of-class-xt-war)
// For the moment, as a quick (and dirty) hack under windows,
// the const variable LogLevels::Levels is defined twice :
// inside "BabylonCpp" lib *AND* inside "inspector" lib.

+++ src/Inspector/src/inspector/components/actiontabs/tabs/statistics_tab_component.cpp
@@ -118,32 +118,32 @@ void StatisticsTabComponent::render()
                        / sceneInstrumentation->frameTimeCounter().current()));
     ValueLineComponent::render(
       "Meshes selection",
-      sceneInstrumentation->activeMeshesEvaluationTimeCounter().current(),
+      static_cast<float>(sceneInstrumentation->activeMeshesEvaluationTimeCounter().current()),
       std::nullopt, std::nullopt, "ms");
     ValueLineComponent::render(
       "Render targets",
-      sceneInstrumentation->renderTargetsRenderTimeCounter().current(),
+      static_cast<float>(sceneInstrumentation->renderTargetsRenderTimeCounter().current()),
       std::nullopt, std::nullopt, "ms");
     ValueLineComponent::render(
-      "Particles", sceneInstrumentation->particlesRenderTimeCounter().current(),
+      "Particles", static_cast<float>(sceneInstrumentation->particlesRenderTimeCounter().current()),
       std::nullopt, std::nullopt, "ms");
     ValueLineComponent::render(
-      "Sprites", sceneInstrumentation->spritesRenderTimeCounter().current(),
+      "Sprites", static_cast<float>(sceneInstrumentation->spritesRenderTimeCounter().current()),
       std::nullopt, std::nullopt, "ms");
     ValueLineComponent::render(
-      "Animations", sceneInstrumentation->animationsTimeCounter().current(),
+      "Animations", static_cast<float>(sceneInstrumentation->animationsTimeCounter().current()),
       std::nullopt, std::nullopt, "ms");
     ValueLineComponent::render(
-      "Physics", sceneInstrumentation->physicsTimeCounter().current(),
+      "Physics", static_cast<float>(sceneInstrumentation->physicsTimeCounter().current()),
       std::nullopt, std::nullopt, "ms");
     ValueLineComponent::render(
-      "Render", sceneInstrumentation->renderTimeCounter().current(),
+      "Render", static_cast<float>(sceneInstrumentation->renderTimeCounter().current()),
       std::nullopt, std::nullopt, "ms");
     ValueLineComponent::render(
-      "Frame total", sceneInstrumentation->frameTimeCounter().current(),
+      "Frame total", static_cast<float>(sceneInstrumentation->frameTimeCounter().current()),
       std::nullopt, std::nullopt, "ms");
     ValueLineComponent::render(
-      "Inter-frame", sceneInstrumentation->interFrameTimeCounter().current(),
+      "Inter-frame", static_cast<float>(sceneInstrumentation->interFrameTimeCounter().current()),
       std::nullopt, std::nullopt, "ms");
     ValueLineComponent::render(
       "GPU Frame time",
